修复数据库，修复逆向工程修复Mapper问题

+++ src/main/java/life/my/community/controller/PublishController.java
@@ -40,7 +40,7 @@ public class PublishController {
             @RequestParam(value = "title",required = false) String title,
             @RequestParam(value = "description", required = false) String description,
             @RequestParam(value = "tag",required = false) String tag,
-            @RequestParam(value = "id",required = false) Integer id,
+            @RequestParam(value = "id",required = false) Long id,
             HttpServletRequest request,
             Model model) {
         model.addAttribute("title",title);
